Reimplement template
Fun fact: it was successfully generated using Python-Markdown (including
inline access to the models).
The only required correction was remove of the `<p></p>` tag, and only
because I wanted to be fully consistent with the code in the Django's
tutorial.
The tag could be present, and template still would be correct
syntax-wise and the view would met expectations.